Update major design and linke BE

+++ src/modules/offers/components/offer.form.js
@@ -1,4 +1,5 @@
 import React,{useEffect, useState} from "react";
+import * as yup from 'yup';
 import DatePicker from "react-datepicker";
 import { FileUpload } from "../../../components/file-upload";
 import { OfferModel } from "../../../core/models/offer.model";
@@ -11,20 +12,34 @@ import fileService from '../../../core/service/fileService';
 import { CategorySelector } from "../../category/components/categorySelector";
 import authService from '../../auth/auth-service';
 import {useAuth} from '../../../core/hooks/useAuth';
+import brandService from '../../brand/service';
+import {useFormValidation} from '../../../core/hooks/useFormValidation';
+import {WrapErrorMessage, ErrorMessage, getErrorClass} from '../../../components/error-message';
+
+const OFFER_FORM_VALIDATION_SCHEMA = yup.object({
+    Title : yup.string().required('required'),
+    StartDate : yup.string().required('required'),
+    EndDate: yup.string().required('required'),
+    CategoryID : yup.number().min(1, 'required').required('required')
+});
 
 export const OfferForm = (props)=>{
     const [offer, setOffer] = useState(new OfferModel());
     const [offerPictureFile, setOfferPictureFile] = useState(null);
     const [offerPicturePreview, setOfferPicturePreview] = useState();
     const isEdit = !isUndefinedNullOrEmpty(props.selectedId);
-    const {user} = useAuth();
+    const {user, isLogin, isBrandUser} = useAuth();
+    const {isValid, errors, onValidate} = useFormValidation(OFFER_FORM_VALIDATION_SCHEMA);
+    const Error = WrapErrorMessage(ErrorMessage, errors);
+    const BrandID = isLogin && isBrandUser() ? user.BrandID : 0;
     
     useEffect(()=>{
-        isEdit && fetchOffer();
+        isEdit ? fetchOffer() : setDefaultBrandCategory() ;
     }, []);
 
     const onChange = (value = {}) => {
         const constructValues = {...offer, ...value};
+        onValidate(constructValues);
         setOffer(constructValues);
     } 
 
@@ -47,26 +62,37 @@ export const OfferForm = (props)=>{
         setOfferPictureFile(null);
     };
 
-    const onSave = ()=>{
-
-        (isEdit ? offerService.updateOffer(offer) : authService.getBrandIdByUserId(user.UserID).then((data)=> offerService.createOffer({...offer, BrandID: data.BrandID}))).then((data)=>{
-
-            if(offerPictureFile){
-                const fileToInsert = offerPictureFile;
-                fileToInsert.append('ReferenceID', isEdit ? offer.OfferID : data.id);
-
-                (isEdit ? fileService.removeFilesByModuleAndRefAndFileUsage(MODULE_TYPE.Offer, offer.OfferID, FILE_USAGE_TYPE.OfferPostImage) : Promise.resolve()).then(()=>{
-                    fileService.createFile(fileToInsert).then((data)=>{
-                        props.onClose();
-                    });
-                })
-                
+    const setDefaultBrandCategory = ()=>{
+        brandService.getBrand(BrandID).then((data)=>{
+            if(!isUndefinedNullOrEmpty(data)){
+                onChange({CategoryID : data.DefaultCategoryID});
             }
-            else{
-                props.onClose();
-            }
-
         });
+    }
+
+    const onSave = ()=>{
+        onValidate(offer).then((isFormValid)=>{
+            if(isFormValid){
+                (isEdit ? offerService.updateOffer(offer) : offerService.createOffer({...offer, BrandID})).then((data)=>{
+
+                    if(offerPictureFile){
+                        const fileToInsert = offerPictureFile;
+                        fileToInsert.append('ReferenceID', isEdit ? offer.OfferID : data.id);
+
+                        (isEdit ? fileService.removeFilesByModuleAndRefAndFileUsage(MODULE_TYPE.Offer, offer.OfferID, FILE_USAGE_TYPE.OfferPostImage) : Promise.resolve()).then(()=>{
+                            fileService.createFile(fileToInsert).then((data)=>{
+                                props.onClose();
+                            });
+                        })
+                        
+                    }
+                    else{
+                        props.onClose();
+                    }
+
+                });
+            }
+        })
         
     }
 
@@ -83,7 +109,8 @@ export const OfferForm = (props)=>{
                 <div class="row form-group">
                     <div class="col-md-6">
                         <label class="text-black" >Title</label>
-                        <input type="text" class="form-control" value={offer.Title} onChange={(e)=> onChange({Title: e.target.value})}/>
+                        <input type="text" class={`form-control ${getErrorClass('Title',errors)}`} value={offer.Title} onChange={(e)=> onChange({Title: e.target.value})}/>
+                        <Error property="Title"/>
                     </div>
                     <div class="col-md-6">
                         <label class="text-black" >PromoCode</label>
@@ -95,17 +122,20 @@ export const OfferForm = (props)=>{
                     <div class="col-md-6 d-flex flex-column">
                         <label class="text-black" >Start Date</label>
                         <DatePicker selected={offer.StartDate} className="form-control" onChange={(date) => onChange({StartDate: date})}/>
+                        <Error property="StartDate"/>
                     </div>
                     <div class="col-md-6 d-flex flex-column">
                         <label class="text-black" >Expired Date</label>
                         <DatePicker selected={offer.EndDate} className="form-control" onChange={(date) => onChange({EndDate: date})}/>
+                        <Error property="EndDate"/>
                     </div>
                 </div>
 
                 <div class="row form-group">
                     <div class="col-md-6">
                         <label class="text-black" >Category</label>
-                        <CategorySelector CategoryID={offer.CategoryID} onChange={(e) => onChange({CategoryID: e.target.value})}/>
+                        <CategorySelector CategoryID={offer.CategoryID} onChange={(e) => onChange({CategoryID: e.target.value})} className={`${getErrorClass('CategoryID',errors)}`}/>
+                        <Error property="CategoryID"/>
                     </div>
                     <div class="col-md-6">
                         <label class="text-black" >Tags Eg - (newyear, ramazan)</label>
@@ -138,7 +168,7 @@ export const OfferForm = (props)=>{
                 <div class="row form-group mt-4">
                     <div class="col-md-12 d-flex justify-content-end">
                         <button class="btn btn-secondary mr-2" onClick={props.onClose}>Cancel</button>
-                        <button class="btn btn-primary" onClick={onSave}>Save</button>
+                        <button class="btn btn-primary" disabled={!isValid} onClick={onSave}>Save</button>
                     </div>
                 </div>
             </div>
